fix(db_sort_players_alphabetically) : was badly updating database
refactor : move displays parts of the code in db_show_all_players function to view.py

+++ controllers/main_controller.py
@@ -2,7 +2,6 @@ from models.player import Player
 from models.tournament import Tournament
 from views.view import View
 
-from prettytable import PrettyTable
 from tinydb import TinyDB
 
 db = TinyDB("data/db.json")
@@ -34,35 +33,21 @@ class Controller:
         while True:
             choice = self.view.players_menu()
             if choice == "1":
-                self.show_all_players()
+                self.db_show_all_players()
             elif choice == "2":
                 self.db_add_players()
             elif choice == "3":
-                self.edit_player()
+                self.db_edit_player()
             elif choice == "0":
                 return
 
-    def show_all_players(self):
+    def db_show_all_players(self):
         self.db_sort_players_alphabetically()
 
         players_table = db.table("players")
         players = players_table.all()
 
-        table = PrettyTable()
-        table.title = "<<<REGISTERED PLAYERS>>>"
-        table.field_names = [
-            "National Chess ID", "Name", "Surname", "Birthday"
-        ]
-
-        for player in players:
-            table.add_row([
-                player["national_chess_id"],
-                player["name"],
-                player["surname"],
-                player["birthday"]
-            ])
-
-        self.view.print(table)
+        self.view.show_all_players(players)
 
     def db_add_players(self):
         while True:
@@ -90,21 +75,25 @@ class Controller:
         Sort players alphabetically by name and surname in the players table
         in the database, and update their keys to match their index.
         """
-
         players_table = db.table("players")
+
         players = players_table.all()
+        if not players:
+            return
 
         sorted_players = sorted(
-            players, key=lambda x: (x["name"].lower(), x["surname"].lower())
+            players, key=lambda x: (
+                x["name"].lower(), x["surname"].lower()
+            )
         )
-
-        sorted_players_dict = {}
-        for index, player in enumerate(sorted_players, start=1):
-            sorted_players_dict[str(index)] = player
 
         players_table.truncate()
 
-        players_table.insert({"players": sorted_players_dict})
+        for player in sorted_players:
+            players_table.insert(player)
+
+    def db_edit_player(self):
+        pass
 
     def create_tournament(self):
         (name, location, description, number_of_rounds) = (
